Implemented Analytics tests and updated proj layout
